+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ---
 
+#### 10.12.2018
+
+* [MainActivity] Added a bottom bar to this activity as well
+* [MainActivity] Fixed a minor issue with SearchView focusability 
+
 #### 9.12.2018
 
-* [MainActivity] Fixed a minor issue with SearchView focusability 
 * [MainActivity] Proper Search Widget instead of an EditText
 
 #### 8.12.2018